[ES|QL] Cleanup after moving commands related code to registry
