+++ src/platform/packages/shared/kbn-esql-ast/src/commands_registry/README.md
