+++ src/platform/packages/shared/kbn-esql-ast/src/definitions/utils/validation/function.ts
