+++ src/platform/packages/shared/kbn-esql-ast/src/definitions/utils/validation/index.ts
@@ -6,5 +6,5 @@
  * your election, the "Elastic License 2.0", the "GNU Affero General Public
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
-export { validateSources } from './sources_validation';
-export { validateCommandArguments } from './validate_command_arguments';
+export { validateSources } from './sources';
+export { validateCommandArguments } from './commands';
