--- src/platform/packages/shared/kbn-esql-validation-autocomplete/src/autocomplete/factories.ts
@@ -1,27 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
- * Public License v 1"; you may not use this file except in compliance with, at
- * your election, the "Elastic License 2.0", the "GNU Affero General Public
- * License v3.0 only", or the "Server Side Public License, v 1".
- */
-import { i18n } from '@kbn/i18n';
-import { TRIGGER_SUGGESTION_COMMAND } from '@kbn/esql-ast';
-import { ISuggestionItem } from '@kbn/esql-ast/src/commands_registry/types';
-
-export const buildValueDefinitions = (
-  values: string[],
-  options?: { advanceCursorAndOpenSuggestions?: boolean; addComma?: boolean }
-): ISuggestionItem[] =>
-  values.map((value) => ({
-    label: `"${value}"`,
-    text: `"${value}"${options?.addComma ? ',' : ''}${
-      options?.advanceCursorAndOpenSuggestions ? ' ' : ''
-    }`,
-    detail: i18n.translate('kbn-esql-validation-autocomplete.esql.autocomplete.valueDefinition', {
-      defaultMessage: 'Literal value',
-    }),
-    kind: 'Value',
-    command: options?.advanceCursorAndOpenSuggestions ? TRIGGER_SUGGESTION_COMMAND : undefined,
-  }));
